service/waltrac/requirements.txt
aiocoap
click
paho-mqtt